:wrench: feat. Day2 - Day3 관계, 기능접근 설정 - Endpoint 마다 테스트 완료 - WebSecurityConfig 에서 url 별 접근 지정 - Entity 끼리 ManyToOne, OneToMany 로 Join - 전체적으로 요구사항에 따라 관계 설정 - dto 도 변동에 따라 수정 - Service 코드들 전체적으로 이전에 writer & password 로 인증하였던 코드에서 => 발급한 토큰으로 인증하도록 전체 수정 (요구사항에 따라 단순 조회 등은 인증없이 조회가능)

+++ src/main/java/com/example/miniProject/service/CommentService.java
@@ -1,6 +1,8 @@
 package com.example.miniProject.service;
 
-import com.example.miniProject.dto.RequestUserDto;
+import com.example.miniProject.auth.entity.UserEntity;
+import com.example.miniProject.auth.jwt.JwtTokenUtils;
+import com.example.miniProject.auth.repository.UserRepository;
 import com.example.miniProject.dto.comment.CommentDto;
 import com.example.miniProject.dto.ResponseDto;
 import com.example.miniProject.dto.comment.RequestReplyDto;
@@ -9,12 +11,14 @@ import com.example.miniProject.entity.CommentEntity;
 import com.example.miniProject.entity.SalesItemEntity;
 import com.example.miniProject.repository.CommentRepository;
 import com.example.miniProject.repository.SalesItemRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
@@ -29,13 +33,16 @@ public class CommentService {
     private final SalesItemRepository salesItemRepository;
 
     public ResponseDto createComment(Long itemId, CommentDto dto) {
-        if (!salesItemRepository.existsById(itemId))
+        UserEntity userEntity = getUser();
+        Optional<SalesItemEntity> optionalSalesItem = salesItemRepository.findById(itemId);
+        if (optionalSalesItem.isEmpty())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         CommentEntity commentEntity = new CommentEntity();
         commentEntity.setWriter(dto.getWriter());
         commentEntity.setPassword(dto.getPassword());
+        commentEntity.setItem(optionalSalesItem.get());
         commentEntity.setContent(dto.getContent());
-        commentEntity.setItemId(itemId);
+        commentEntity.setUser(userEntity);
 
         commentRepository.save(commentEntity);
 
@@ -55,13 +62,18 @@ public class CommentService {
     }
 
     public ResponseDto updateComment(Long itemId, Long commentId, CommentDto dto) {
+        UserEntity userEntity = getUser();
         if (!salesItemRepository.existsById(itemId))
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         Optional<CommentEntity> optionalCommentEntity = commentRepository.findById(commentId);
         if (optionalCommentEntity.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
 
         CommentEntity targetEntity = optionalCommentEntity.get();
-        validateWriterAndPassword(targetEntity, dto.getWriter(), dto.getPassword());
+        if (!itemId.equals(targetEntity.getItem().getId()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+        if (!targetEntity.getUser().getId().equals(userEntity.getId()))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
 
         targetEntity.setContent(dto.getContent());
         commentRepository.save(targetEntity);
@@ -70,38 +82,60 @@ public class CommentService {
     }
 
     public ResponseDto createReply(Long itemId, Long commentId, RequestReplyDto dto) {
+        UserEntity userEntity = getUser();
         Optional<SalesItemEntity> salesItemEntity = salesItemRepository.findById(itemId);
         if (salesItemEntity.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
 
         Optional<CommentEntity> commentEntity = commentRepository.findById(commentId);
         if (commentEntity.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
 
-        if (!salesItemEntity.get().getWriter().equals(dto.getWriter())
-            || !salesItemEntity.get().getPassword().equals(dto.getPassword())) {
+        SalesItemEntity itemEntity = salesItemEntity.get();
+        CommentEntity comment = commentEntity.get();
+        if (!itemId.equals(comment.getItem().getId()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        if (!itemEntity.getUser().getId().equals(userEntity.getId()))
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
-        }
-        CommentEntity entity = commentEntity.get();
-        entity.setReply(dto.getReply());
-        commentRepository.save(entity);
+        comment.setReply(dto.getReply());
+        commentRepository.save(comment);
         return new ResponseDto("댓글에 답변이 추가되었습니다.");
     }
 
-    public ResponseDto deleteComment(Long itemId, Long commentId, RequestUserDto dto) {
+    public ResponseDto deleteComment(Long itemId, Long commentId) {
+        UserEntity userEntity = getUser();
         if (!salesItemRepository.existsById(itemId))
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         Optional<CommentEntity> optionalCommentEntity = commentRepository.findById(commentId);
         if (optionalCommentEntity.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND);
 
         CommentEntity targetEntity = optionalCommentEntity.get();
-        validateWriterAndPassword(targetEntity, dto.getWriter(), dto.getPassword());
-
+        if (!itemId.equals(targetEntity.getItem().getId()))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        if (!targetEntity.getUser().getId().equals(userEntity.getId()))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         commentRepository.delete(targetEntity);
         return new ResponseDto("댓글을 삭제했습니다.");
     }
 
-    private void validateWriterAndPassword(CommentEntity commentEntity, String writer, String password) {
-        if (!commentEntity.getPassword().equals(password) || !commentEntity.getWriter().equals(writer)) {
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+    private final HttpServletRequest request;
+    private final JwtTokenUtils jwtTokenUtils;
+    private final UserRepository userRepository;
+
+    private UserEntity getUser() {
+        String token = extractTokenFromHeader(request.getHeader(HttpHeaders.AUTHORIZATION));
+        if (jwtTokenUtils.validate(token)) {
+            String username = jwtTokenUtils.parseClaims(token).getSubject();
+            return userRepository.findByUsername(username)
+                    .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        } else {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "유효하지 않은 토큰입니다");
         }
     }
+
+    private String extractTokenFromHeader(String authHeader) {
+        if (authHeader != null && authHeader.startsWith("Bearer ")) {
+            return authHeader.split(" ")[1];
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "토큰의 형식이 잘못되었습니다");
+    }
+
 }
